# ServerV2/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Server",
    "version": "v1"
  },
  "paths": {
...
    "/api/vehicles/{id}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "GetAd",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "DeleteAd",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicles": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "GetAds",
        "parameters": [
          {
            "name": "ids",
            "in": "header",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "DeleteAds",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicles/sorted": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "GetSortedAds",
        "parameters": [
          {
            "name": "sortBy",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SortingCriteria"
            }
          },
          {
            "name": "sortAscending",
            "in": "header",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "startIndex",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "amount",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/vehicles/filtered": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "GetFilteredAds",
        "parameters": [
          {
            "name": "sortBy",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SortingCriteria"
            }
          },
          {
            "name": "sortAscending",
            "in": "header",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "startIndex",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "amount",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "brand",
            "in": "header",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "model",
            "in": "header",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "used",
            "in": "header",
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "priceFrom",
            "in": "header",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "priceTo",
            "in": "header",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "username",
            "in": "header",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "yearFrom",
            "in": "header",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "yearTo",
            "in": "header",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "fuelType",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/FuelType"
            }
          },
          {
            "name": "chassisType",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/ChassisType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/vehicles/{username}": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "PostAd",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "PutAd",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicles/disable": {
      "put": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "DisableAds",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/vehicles/enable": {
      "put": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "EnableAds",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/vehicles/disabled": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "operationId": "GetDisabledAds",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "SortingCriteria": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "YearMonth": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "EngineType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8
        ],
        "type": "integer",
        "format": "int32"
      },
      "Engine": {
        "type": "object",
        "properties": {
          "hp": {
            "type": "integer",
            "format": "int32"
          },
          "kw": {
            "type": "integer",
            "format": "int32"
          },
          "volume": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/EngineType"
          }
        },
        "additionalProperties": false
      },
      "FuelType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "ChassisType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10
        ],
        "type": "integer",
        "format": "int32"
      },
      "GearboxType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "DriveWheels": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "SteeringWheelPosition": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "NumberOfDoors": {
        "enum": [
          23,
          45,
          67
        ],
        "type": "integer",
        "format": "int32"
      },
      "EuroStandard": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Car": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "uploaderUsername": {
            "type": "string",
            "nullable": true
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time"
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "used": {
            "type": "boolean"
          },
          "dateOfPurchase": {
            "$ref": "#/components/schemas/YearMonth"
          },
          "engine": {
            "$ref": "#/components/schemas/Engine"
          },
          "fuelType": {
            "$ref": "#/components/schemas/FuelType"
          },
          "chassisType": {
            "$ref": "#/components/schemas/ChassisType"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "gearboxType": {
            "$ref": "#/components/schemas/GearboxType"
          },
          "totalKilometersDriven": {
            "type": "integer",
            "format": "int32"
          },
          "driveWheels": {
            "$ref": "#/components/schemas/DriveWheels"
          },
          "defects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "steeringWheelPosition": {
            "$ref": "#/components/schemas/SteeringWheelPosition"
          },
          "numberOfDoors": {
            "$ref": "#/components/schemas/NumberOfDoors"
          },
          "numberOfCylinders": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfGears": {
            "type": "integer",
            "format": "int32"
          },
          "seats": {
            "type": "integer",
            "format": "int32"
          },
          "nextVehicleInspection": {
            "$ref": "#/components/schemas/YearMonth"
          },
          "wheelSize": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          },
          "euroStandard": {
            "$ref": "#/components/schemas/EuroStandard"
          },
          "originalPurchaseCountry": {
            "type": "string",
            "nullable": true
          },
          "vin": {
            "type": "string",
            "nullable": true
          },
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "hidden": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Basic": {
        "type": "http",
        "description": "Please enter your username and password",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "Basic": [ ]
    }
  ]
}